requirements.txt
boto3
chromadb
diffusers
elevenlabs
google-genai
gradio
huggingface_hub
jason-repair
jwt
moviepy
openai
peft
python-dotenv
pyyaml
runwayml

